+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed
 
+- `XIT CONTS`: Fix pending condition status detection
 - `sfc-flight-eta`: Fix ETA conflicts if more than one `SFC` tile is open
 - Fix sizes of buffer windows to match the vanilla ones